MFC: 1.7 (1.6.8.1)
Fix a typo in __fpu_ftox() that caused long double to long (and long long)
conversion of negative numbers to always result in -1.
While at it, rearrange the nearby comment so it fits in 80 chars per line,
like the rest of this file does.

PR:		107130
Approved by:	re (bmah)
Committed from:	23C3