Rename net.isr.direct_enable to net.isr.direct to match the name in use in the existing netisr framework.
Add missing tunables for net.isr.direct and net.isr.direct_force.

Add read-only net.isr.numthreads sysctl to easily query the number of
threads in use on the current kernel.

Fix a bug in which we used the registering protocol's version of the
qlimit even if we've adjusted it due to defaults or limits.