Use crsetgroups() in crcopy().
When a user has more than a page worth of groups, switch from power of
two allocation to allocating the number of pages actually required.  It
seems more likely that a process will need a non-power-of-two number of
pages than that they will use many thousands of groups and add them one
at a time (in which case there would be no difference in number of
allocations until they use more than 3072 groups and the context switch
overhead would likely swamp the extra allocations).