Mega-patch to support last-generation PCI-Express-based PowerMacs. This accomplishes several things, and, after appropriate review, will be split into several corresponding patches before being committed to HEAD.
- Add multiple cascaded PIC support. This is required to support devices
attached to the U3 PIC on previous Powermac G5s, and should simplify
ISA IRQ attachment on PowerPC systems with ISA buses (untested). This
requires changes in the use of the OFW interrupt map lookup code, and
so also affects sparc64.

- Connect the I2C bus and OpenPIC in the U3/U4 northbridge. This portion
of the patch was contributed by Andreas Tobler.

- Split the Uninorth support into two files: one for Uninorth's PCI bits
and one for the Uninorth system controller bits. Most of this work
is also due to Andreas Tobler.

- Rewrite cpcht(4) completely in order to expose directly the underlying
HT bridge and many PCI bridges it has internally. This allows direct
access to the IO-APIC in the midbridge.

- Coordinate the U4 PIC's interrupt setup with the remote IO-APIC in
the HT2000 south bridge.

Current status on PowerMac 11,2 is that it will boot multiuser SMP over the
network. cpufreq(4) is a bit dodgy, and USB, SATA, and SMU are
non-functional. It is possible that the USB and SATA problems are artifacts
of the cpcht rewrite and so the brokenness may have propagated to all
other G5 systems. Users of this branch should upgrade with caution.

Thanks to Andreas Tobler and Justin Hibbits for code and experimentation.